+++ tests/crawlers/test_cases/spankbang.py
@@ -5,7 +5,7 @@ TEST_CASES = [
         "results": [
             {
                 "url": "re:https://vdownload.+16371508-1080p.mp4\\?.+_tid=16371508",
-                "filename": "WIFE FOR FREE USE . Devil Khloee [9qwc4][1080p].mp4",
+                "filename": "WIFE FOR FREE USE. Devil Khloee [9qwc4][1080p].mp4",
                 "referer": "https://spankbang.com/9qwc4/video/wife+for+free+use+devil+khloee",
                 "album_id": None,
                 "uploaded_at": None,
@@ -13,53 +13,54 @@ TEST_CASES = [
         ],
     },
     {
-        "url": "https://spankbang.com/9ndp4/video/korean+dance",
+        "url": "https://fr.spankbang.com/9lcld/video/3364+kbj+korean",
         "results": [
             {
-                "url": "re:/1/6/16207384-4k.mp4",
-                "filename": "Korean dance [9ndp4][4K].mp4",
+                "url": "re:/1/6/16112641-4k.mp4",
+                "original_filename": "3364 KBJ 벗방 팬방 여캠 딥쓰롯 리액션 저게 다들어가네 ㅋㅋ Korean",
+                "referer": "https://spankbang.com/9lcld/video/3364+kbj+korean",
+            }
+        ],
+    },
+    {
+        "url": "https://jp.spankbang.com/9lcld/video/3364+kbj+korean",
+        "results": [
+            {
+                "url": "re:/1/6/16112641-4k.mp4",
+                "original_filename": "3364 KBJ 벗방 팬방 여캠 딥쓰롯 리액션 저게 다들어가네 ㅋㅋ Korean",
+                "referer": "https://spankbang.com/9lcld/video/3364+kbj+korean",
+            }
+        ],
+    },
+    {
+        "url": "https://de.spankbang.com/profile/fmsh7827/videos",
+        "results": [
+            {
+                "url": "re:/1/4/14308081-480p.mp4",
+                "filename": "kbj1561 [8io6p][480p].mp4",
                 "debrid_url": None,
-                "original_filename": "Korean dance",
-                "referer": "https://spankbang.com/9ndp4/video/korean+dance",
+                "original_filename": "kbj1561",
+                "referer": "https://spankbang.com/8io6p/video/kbj1561",
                 "album_id": None,
                 "uploaded_at": None,
-                "download_folder": "re:Loose Files (SpankBang)",
+                "download_folder": "re:fmsh7827 [user] (SpankBang)",
             }
         ],
+        "count": 1,
     },
     {
-        "url": "https://fr.spankbang.com/9ndp4/video/korean+dance",
+        "url": "https://spankbang.com/a0cyk-tsmnwz/playlist/korbj",
         "results": [
             {
-                "url": "re:/1/6/16207384-4k.mp4",
-                "original_filename": "Korean dance",
-                "referer": "https://spankbang.com/9ndp4/video/korean+dance",
-            }
-        ],
-    },
-    {
-        "url": "https://jp.spankbang.com/9ndp4/video/korean+dance",
-        "results": [
-            {
-                "url": "re:/1/6/16207384-4k.mp4",
-                "original_filename": "Korean dance",
-                "referer": "https://spankbang.com/9ndp4/video/korean+dance",
-            }
-        ],
-    },
-    {
-        "url": "https://spankbang.com/a0cyk-rb6ofs/playlist/korbj",
-        "results": [
-            {
-                "url": "re:/1/6/16207384-4k.mp4",
-                "original_filename": "Korean dance",
-                "referer": "https://spankbang.com/9ndp4/video/korean+dance",
+                "url": "re:/1/4/14234633-480p.mp4",
+                "original_filename": "korean",
+                "referer": "https://spankbang.com/8h3ih/video/korean",
             }
         ],
     },
     {
         "url": "https://de.spankbang.com/a0cyk/playlist/korbj",
         "results": [{"url": "ANY", "download_folder": "re:Korbj [playlist] (SpankBang)"}],
-        "count": range(180, 220),
+        "count": range(240, 260),
     },
 ]