+++ CHANGELOG.md
@@ -29,6 +29,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove hardcoded limit of 1 concurrent download per server (Bunkr)
 - Do not dedupe files that were not downloaded in the current session
 
+### Fixed
+
+- No videos detected on profiles, playlists or search results (Spankbang)
+
 ## [10.0.2] - 2026-07-05
 
 ### Fixed